Split network DI module to debug and release versions. Init Stetho only for the Debug one
