+++ client/src/debug/java/io/getstream/chat/android/client/di/ChatModule.kt
@@ -1,0 +1,26 @@
+package io.getstream.chat.android.client.di
+
+import android.content.Context
+import com.facebook.stetho.Stetho
+import com.facebook.stetho.okhttp3.StethoInterceptor
+import io.getstream.chat.android.client.api.ChatClientConfig
+import io.getstream.chat.android.client.parser.ChatParser
+import okhttp3.OkHttpClient
+
+internal class ChatModule(appContext: Context, config: ChatClientConfig) : BaseChatModule(appContext, config) {
+
+    init {
+        Stetho.initializeWithDefaults(appContext)
+    }
+
+    override fun clientBuilder(
+        connectTimeout: Long,
+        writeTimeout: Long,
+        readTimeout: Long,
+        config: ChatClientConfig,
+        parser: ChatParser
+    ): OkHttpClient.Builder {
+        return super.clientBuilder(connectTimeout, writeTimeout, readTimeout, config, parser)
+            .addNetworkInterceptor(StethoInterceptor())
+    }
+}
