+++ client/src/release/java/io.getstream.chat.android.client.di/ChatModule.kt
@@ -1,0 +1,6 @@
+package io.getstream.chat.android.client.di
+
+import android.content.Context
+import io.getstream.chat.android.client.api.ChatClientConfig
+
+internal class ChatModule(appContext: Context, config: ChatClientConfig) : BaseChatModule(appContext, config)
